+++ features/step_definitions/posts_steps.rb
@@ -32,6 +32,10 @@ Then /^I should not be able to submit the publisher$/ do
   expect(publisher_submittable?).to be false
 end
 
+Then /^I should see "([^"]*)" in the publisher$/ do |text|
+  expect(page).to have_field("status_message[text]", with: text)
+end
+
 Given /^I have a limited post with text "([^\"]*)" in the aspect "([^"]*)"$/ do |text, aspect_name|
   @me.post :status_message, text: text, to: @me.aspects.where(name: aspect_name).first.id
 end
@@ -112,10 +116,6 @@ end
 
 When /^I append "([^"]*)" to the publisher$/ do |text|
   append_to_publisher(text)
-end
-
-When /^I append "([^"]*)" to the mobile publisher$/ do |text|
-  append_to_publisher(text, '#status_message_text')
 end
 
 When /^I attach "([^"]*)" to the publisher$/ do |path|